requirements.txt
flask>=2.3.0
tensorflow>=2.12.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.8.0
flask-cors>=4.0.0 